completely rewrite using opencv matrix

+++ fisheye_cam.hpp
@@ -2,7 +2,7 @@
 * @Author: BlahGeek
 * @Date:   2015-10-14
 * @Last Modified by:   BlahGeek
-* @Last Modified time: 2015-10-14
+* @Last Modified time: 2015-10-18
 */
 
 #ifndef VR_LIBMAP_FISHEYE_H
@@ -16,8 +16,8 @@ class FisheyeCamera: public PinholeCamera {
 public:
     using PinholeCamera::PinholeCamera;
 protected:
-    void _project(std::vector<cv::Point3f> & objectPoints,
-                  std::vector<cv::Point2f> & imagePoints) override;
+    void _project(std::vector<cv::Point3d> & objectPoints,
+                  std::vector<cv::Point2d> & imagePoints) override;
 };
 
 }
